+++ MatrixSDK/Crypto/CrossSigning/MXCrossSigning_Private.h
@@ -41,10 +41,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (BOOL)isSecretValid:(NSString*)secret forPublicKeys:(NSString*)keys;
 
-- (nullable NSString*)secretIdFromKeyType:(NSString*)keyType;
-
-- (void)signDevice:(MXDeviceInfo*)device
-           success:(void (^)(void))success
+- (void)signObject:(NSDictionary*)object withKeyType:(NSString*)keyType
+           success:(void (^)(NSDictionary *signedObject))success
            failure:(void (^)(NSError *error))failure;
 
 @end