+++ MatrixSDK/MXSession.m
@@ -203,6 +203,8 @@ typedef void (^MXOnResumeDone)(void);
 
 @property (nonatomic, strong) id<MXSyncResponseStore> initialSyncResponseCache;
 
+@property (atomic, copy, readwrite) NSDictionary<NSString*, NSArray<NSString*>*> *directRooms;
+
 @end
 
 @implementation MXSession
@@ -301,7 +303,7 @@ typedef void (^MXOnResumeDone)(void);
 {
     if (_state != state)
     {
-        MXLogDebug(@"[MXSession] setState: %@ (was %@)", @(state), @(_state));
+        MXLogDebug(@"[MXSession] setState: %@ (was %@)", [MXTools readableSessionState:state], [MXTools readableSessionState:_state]);
         
         _state = state;
 
@@ -658,6 +660,8 @@ typedef void (^MXOnResumeDone)(void);
                     [self.crypto handleDeviceOneTimeKeysCount:syncResponse.deviceOneTimeKeysCount];
                 }
                 
+                [self.crypto handleDeviceUnusedFallbackKeys:syncResponse.unusedFallbackKeys];
+                
                 // Tell the crypto module to do its processing
                 [self.crypto onSyncCompleted:self.store.eventStreamToken
                                nextSyncToken:syncResponse.nextBatch
@@ -667,6 +671,11 @@ typedef void (^MXOnResumeDone)(void);
             // Update live event stream token
             MXLogDebug(@"[MXSession] Next sync token: %@", syncResponse.nextBatch);
             self.store.eventStreamToken = syncResponse.nextBatch;
+            
+            if (self.spaceService.needsUpdate || syncResponse.rooms.join.count || syncResponse.rooms.invite.count || syncResponse.rooms.leave.count || syncResponse.toDevice.events.count)
+            {
+                [self.spaceService buildGraphWith:self.rooms];
+            }
             
             if (completion)
             {
@@ -910,11 +919,19 @@ typedef void (^MXOnResumeDone)(void);
     return _store.syncFilterId;
 }
 
+- (BOOL)isPauseable
+{
+    return _state == MXSessionStateSyncInProgress
+        || _state == MXSessionStateRunning
+        || _state == MXSessionStateBackgroundSyncInProgress
+        || _state == MXSessionStatePauseRequested;
+}
+
 - (void)pause
 {
-    MXLogDebug(@"[MXSession] pause the event stream in state %tu", _state);
+    MXLogDebug(@"[MXSession] pause the event stream in state %@", [MXTools readableSessionState:_state]);
 
-    if (_state == MXSessionStateRunning || _state == MXSessionStateBackgroundSyncInProgress || _state == MXSessionStatePauseRequested)
+    if (self.isPauseable)
     {
         // Check that none required the session to keep running even if the app goes in
         // background
@@ -972,7 +989,7 @@ typedef void (^MXOnResumeDone)(void);
 
 - (void)_resume:(void (^)(void))resumeDone
 {
-    MXLogDebug(@"[MXSession] _resume: resume the event stream from state %tu", _state);
+    MXLogDebug(@"[MXSession] _resume: resume the event stream from state %@", [MXTools readableSessionState:_state]);
     
     if (self.backgroundTask.isRunning)
     {
@@ -1021,7 +1038,7 @@ typedef void (^MXOnResumeDone)(void);
     {
         if (!ignoreSessionState && MXSessionStatePaused != _state)
         {
-            MXLogDebug(@"[MXSession] background Sync cannot be done in the current state %tu", _state);
+            MXLogDebug(@"[MXSession] background Sync cannot be done in the current state: %@", [MXTools readableSessionState:_state]);
             dispatch_async(dispatch_get_main_queue(), ^{
                 backgroundSyncfails(nil);
             });
@@ -1145,6 +1162,9 @@ typedef void (^MXOnResumeDone)(void);
         [self.backgroundTask stop];
         self.backgroundTask = nil;
     }
+    
+    // Clear spaces
+    [self.spaceService close];
 
     _myUser = nil;
     mediaManager = nil;
@@ -1246,7 +1266,7 @@ typedef void (^MXOnResumeDone)(void);
 {
     _preventPauseCount = preventPauseCount;
 
-    MXLogDebug(@"[MXSession] setPreventPauseCount: %tu. MXSession state: %tu", _preventPauseCount, _state);
+    MXLogDebug(@"[MXSession] setPreventPauseCount: %tu. MXSession state: %@", _preventPauseCount, [MXTools readableSessionState:_state]);
 
     if (_preventPauseCount == 0)
     {
@@ -1552,18 +1572,13 @@ typedef void (^MXOnResumeDone)(void);
         if ([error.domain isEqualToString:NSURLErrorDomain]
             && code == kCFURLErrorCancelled)
         {
-            MXLogDebug(@"[MXSession] The connection has been cancelled in state %@", @(_state));
+            MXLogDebug(@"[MXSession] The connection has been cancelled in state: %@", [MXTools readableSessionState:_state]);
 
-            if (_state == MXSessionStateSyncInProgress)
-            {
-                // This happens when the SDK cannot make any more requests because the app is in background
-                // and the background task is expired or going to expire.
-                // The app should have paused the SDK before but it did not. So, pause the SDK ourselves.
-                // Note that we need to come back to MXSessionStatePauseRequested in order to be able to pause.
-                MXLogDebug(@"[MXSession] -> Go to pause");
-                [self setState:MXSessionStatePauseRequested];
-                [self pause];
-            }
+            // This happens when the SDK cannot make any more requests because the app is in background
+            // and the background task is expired or going to expire.
+            // The app should have paused the SDK before but it did not. So, pause the SDK ourselves.
+            MXLogDebug(@"[MXSession] -> Go to pause");
+            [self pause];
         }
         else if ([error.domain isEqualToString:NSURLErrorDomain]
                  && code == kCFURLErrorTimedOut && serverTimeout == 0)
@@ -1719,14 +1734,16 @@ typedef void (^MXOnResumeDone)(void);
                 NSDictionary<NSString*, NSArray<NSString*>*> *directRooms;
                 MXJSONModelSetDictionary(directRooms, event[@"content"]);
 
-                if (directRooms != _directRooms
-                    && ![directRooms isEqualToDictionary:_directRooms])
+                NSDictionary<NSString*, NSArray<NSString*>*> *tmpDirectRooms = self.directRooms;
+                
+                if (directRooms != tmpDirectRooms
+                    && ![directRooms isEqualToDictionary:tmpDirectRooms])
                 {
                     // Collect previous direct rooms ids
                     NSMutableSet<NSString*> *directRoomIds = [NSMutableSet set];
                     [directRoomIds unionSet:[self directRoomIds]];
 
-                    _directRooms = directRooms;
+                    self.directRooms = directRooms;
 
                     // And collect current ones
                     [directRoomIds unionSet:[self directRoomIds]];
@@ -2009,7 +2026,7 @@ typedef void (^MXOnResumeDone)(void);
         }
         else
         {
-            MXLogDebug(@"[MXSesion] enableCrypto: crypto module will be start later (MXSession.state: %@)", @(_state));
+            MXLogDebug(@"[MXSesion] enableCrypto: crypto module will be start later (MXSession.state: %@)", [MXTools readableSessionState:_state]);
 
             if (success)
             {
@@ -2554,7 +2571,7 @@ typedef void (^MXOnResumeDone)(void);
 - (NSSet<NSString*> *)directRoomIds
 {
     NSMutableSet<NSString*> *roomIds = [NSMutableSet set];
-    for (NSArray *array in _directRooms.allValues)
+    for (NSArray *array in self.directRooms.allValues)
     {
         [roomIds addObjectsFromArray:array];
     }
@@ -2564,16 +2581,16 @@ typedef void (^MXOnResumeDone)(void);
 
 - (NSString *)directUserIdInRoom:(NSString*)roomId
 {
-    NSString *directUserId;
+    __block NSString *directUserId;
 
-    for (NSString *userId in _directRooms)
+    [self.directRooms enumerateKeysAndObjectsUsingBlock:^(NSString * _Nonnull userId, NSArray<NSString *> * _Nonnull roomIds, BOOL * _Nonnull stop)
     {
-        if ([_directRooms[userId] containsObject:roomId])
+        if ([roomIds containsObject:roomId])
         {
             directUserId = userId;
-            break;
+            *stop = YES;
         }
-    }
+    }];
 
     return directUserId;
 }
@@ -2652,10 +2669,12 @@ typedef void (^MXOnResumeDone)(void);
                                                success:(void (^)(void))success
                                                failure:(void (^)(NSError *))failure
 {
+    NSDictionary<NSString*, NSArray<NSString*>*> *tmpDirectRooms = self.directRooms;
+
     // If there is no change, do nothing
-    if (_directRooms == directRooms
-        || [_directRooms isEqualToDictionary:directRooms]
-        || (_directRooms == nil && directRooms.count == 0))
+    if (tmpDirectRooms == directRooms
+        || [tmpDirectRooms isEqualToDictionary:directRooms]
+        || (tmpDirectRooms == nil && directRooms.count == 0))
     {
         if (success)
         {