+++ MatrixSDK/Space/MXSpace.swift
@@ -35,6 +35,7 @@ extension MXSpaceError: CustomNSError {
 }
 
 /// A Matrix space enables to collect rooms together into groups. Such collections of rooms are referred as "spaces" (see https://github.com/matrix-org/matrix-doc/blob/matthew/msc1772/proposals/1772-groups-as-rooms.md).
+@objcMembers
 public class MXSpace: NSObject {
     
     // MARK: - Properties
@@ -52,6 +53,11 @@ public class MXSpace: NSObject {
         return self.room.summary
     }
     
+    public private(set) var childSpaces: [MXSpace] = []
+    public private(set) var childRoomIds: [String] = []
+    public private(set) var otherMembersId: [String] = []
+    public private(set) var membersId: [String] = []
+    
     // MARK: - Setup
     
     public init(room: MXRoom) {
@@ -60,6 +66,42 @@ public class MXSpace: NSObject {
     }
     
     // MARK: - Public
+    
+    /// Update children and members from room states and members
+    /// - Parameters:
+    ///   - completion: A closure called when the operation completes.
+    public func readChildRoomsAndMembers(completion: @escaping () -> Void) {
+        guard let myUserId = room.mxSession.myUserId else {
+            return
+        }
+
+        room.state { [weak self] roomState in
+            roomState?.stateEvents.forEach({ event in
+                if event.eventType == .spaceChild {
+                    self?.childRoomIds.append(event.stateKey)
+                }
+            })
+            
+            self?.room.members { [weak self] response in
+                guard let members = response.value as? MXRoomMembers else {
+                    return
+                }
+
+                var otherMembersId: [String] = []
+                var membersId: [String] = []
+                members.members?.forEach { roomMember in
+                    membersId.append(roomMember.userId)
+                    if roomMember.userId != myUserId {
+                        otherMembersId.append(roomMember.userId)
+                    }
+                }
+                self?.otherMembersId = otherMembersId
+                self?.membersId = membersId
+                
+                completion()
+            }
+        }
+    }
         
     /// Add child space or child room to the current space.
     /// - Parameters:
@@ -107,6 +149,43 @@ public class MXSpace: NSObject {
                                  stateKey: roomId,
                                  completion: completion)
     }
+    
+    /// Update child spaces using the list of spaces
+    /// - Parameters:
+    ///   - spacesPerId: complete list of spaces by space ID
+    public func updateChildSpaces(with spacesPerId: [String: MXSpace]) {
+        self.childRoomIds.forEach { roomId in
+            if let space = spacesPerId[roomId] {
+                self.childSpaces.append(space)
+            }
+        }
+    }
+    
+    /// Update child rooms using the list of direct rooms
+    /// - Parameters:
+    ///   - directRoomsPerMember: complete list of direct rooms by member ID
+    public func updateChildDirectRooms(with directRoomsPerMember: [String : [MXRoom]]) {
+        self.updateChildRooms(from: self, with: directRoomsPerMember)
+    }
+    
+    /// Check if the room identified with an ID is a child of the space
+    /// - Parameters:
+    ///   - roomId: The room id of the potential child room.
+    /// - Returns: `true` if the room identified is a child, `false` atherwise
+    public func isRoomAChild(roomId: String) -> Bool {
+        return childRoomIds.contains(roomId)
+    }
+    
+    // MARK: - Private
+    
+    private func updateChildRooms(from space: MXSpace, with directRoomsPerMember: [String : [MXRoom]]) {
+        space.otherMembersId.forEach { memberId in
+            let rooms = directRoomsPerMember[memberId] ?? []
+            self.childRoomIds.append(contentsOf: rooms.compactMap({ room in
+                return room.roomId
+            }))
+        }
+    }
 }
 
 // MARK: - Objective-C
@@ -124,7 +203,7 @@ extension MXSpace {
     ///   - failure: A closure called  when the operation fails.
     /// - Returns: a `MXHTTPOperation` instance.
     @discardableResult
-    public func addChild(roomId: String,
+    @objc public func addChild(roomId: String,
                          viaServers: [String]?,
                          order: String?,
                          autoJoin: Bool,